refactor: import css-in-js utils from theme
