+++ components/_util/theme/util/index.tsx
@@ -2,6 +2,7 @@
 import { CSSObject } from '@ant-design/cssinjs';
 import type { DerivativeToken } from '..';
 
+export { operationUnit } from './operationUnit';
 export { roundedArrow } from './roundedArrow';
 
 export const resetComponent = (token: DerivativeToken): CSSObject => ({
