+++ components/typography/style/index.tsx
@@ -1,7 +1,6 @@
 // deps-lint-skip-all
-import { FullToken, genComponentStyleHook } from '../../_util/theme';
-import type { GenerateStyle } from '../../_util/theme';
-import { operationUnit } from '../../_util/theme/util/operationUnit';
+import type { FullToken, GenerateStyle } from '../../_util/theme';
+import { genComponentStyleHook, operationUnit } from '../../_util/theme';
 import {
   getTitleStyles,
   getResetStyles,
