+++ site/theme/template/Layout/DynamicTheme/Token.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-array-index-key */
 import * as React from 'react';
 import { Table, Space, TableProps, ConfigProvider, Select, Row, Col, Alert } from 'antd';
-import { statistic } from '../../../../../components/_util/theme/util/statistic';
+import { statistic } from '../../../../../components/_util/theme';
 
 const columns: TableProps<{ name: string; value: any }>['columns'] = [
   {
